+++ apps/api/urls.py
@@ -12,9 +12,9 @@ from .views import (
 
 urlpatterns = [
     path("universities/", UniversityListAPIView.as_view()),
-    path("universities/<slug:slug>/", UniversityDetailAPIView.as_view()),
+    path("universities/<str:slug>/", UniversityDetailAPIView.as_view()),
     path("programs/", ProgramListAPIView.as_view()),
-    path("programs/<slug:slug>/", ProgramDetailAPIView.as_view()),
+    path("programs/<str:slug>/", ProgramDetailAPIView.as_view()),
     path("offerings/", ProgramOfferingListAPIView.as_view()),
     path("faq-categories/", FAQCategoryListAPIView.as_view()),
     path("faqs/", FAQListAPIView.as_view()),